assignment/requirements.txt
streamlit>=1.37.0
pydantic>=2.7.0
openai>=1.40.0
python-dotenv>=1.0.1
Pillow>=10.3.0
pytesseract>=0.3.10
PyMuPDF>=1.24.2
numpy>=1.26.4
rapidfuzz>=3.9.0
json5>=0.9.25
tenacity>=8.3.0 